Replace message list with checkable item list
Replaces the default list of messages created by the used template with
a list of shopilist items. The items shown is a static list that can not
be altered at the moment.

Where possible I took the opportunity to clean some stuff up, i.e.
remove code we won't need, such as the refresher.

Signed-off-by: Niklas Wimmer <nwimmer@student.tgm.ac.at>